Move CodeBlock out of chakra namespace
