+++ reflex/components/datadisplay/__init__.py
@@ -1,6 +1,10 @@
 """Data grid components."""
 
+from .code import CodeBlock
+from .code import LiteralCodeBlockTheme as LiteralCodeBlockTheme
+from .code import LiteralCodeLanguage as LiteralCodeLanguage
 from .dataeditor import DataEditor, DataEditorTheme
 
+code_block = CodeBlock.create
 data_editor = DataEditor.create
 data_editor_theme = DataEditorTheme
